fix(firmware): fix build errors and increase logging resolution
- Fix missing headers and type mismatch warnings in float and surface firmware
- Increase mission logging resolution to 1s and make it configurable
- Consolidate live status metrics into Active Configuration section in UI

+++ lib/pid/depth_pid.c
@@ -34,8 +34,8 @@ void depth_pid_calculate_target_pos(DepthPID *dpid, double current_depth, int ne
      * The 'neutral_adc' baseline is handled internally by the PID's integral term,
      * which is seeded at the start of the mission.
      */
-    double error = current_depth - dpid->target_depth;
-    double pid_output = 0;
+    float error = (float)(current_depth - dpid->target_depth);
+    float pid_output = 0;
     
     pid_update(&dpid->pid, error, &pid_output);
